actualizando repo

+++ clinica-odontologica-tp/clinica-odontologica-fix (1)/clinica-odontologica/src/main/java/com/dh/clinica/service/TurnoService.java
@@ -46,7 +46,7 @@ public class TurnoService {
         return ITurnoRepository.save(turno);
     }
 
-    public Optional<Turno> buscar(Integer id) {
+    public Optional<Turno> buscar(Long id) {
         return ITurnoRepository.findById(id);
     }
